Add restrictions in controllers

+++ src/main/java/edu/example/web/controller/CommentController.java
@@ -5,6 +5,7 @@ import edu.example.dto.comment.CommentResponseDto;
 import edu.example.dto.comment.CreateCommentRequestDto;
 import edu.example.mapper.CommentMapper;
 import edu.example.model.Comment;
+import edu.example.model.Role;
 import edu.example.service.CommentService;
 import edu.example.web.security.UserInfoDetails;
 import io.swagger.v3.oas.annotations.Operation;
@@ -13,6 +14,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -51,7 +53,13 @@ public class CommentController {
             @ApiResponse(responseCode = "404", description = "Comment not found"),
             @ApiResponse(responseCode = "403", description = "Insufficient rights / unauthorised"),
     })
-    public void delete(@PathVariable Long id) {
+    public void delete(@PathVariable Long id, @AuthenticationPrincipal UserInfoDetails userDetails) {
+        if (!(
+                userDetails.getRoles().stream().anyMatch(it -> it.getAuthority().equals(Role.MODERATOR.name())) ||
+                        commentService.getComment(id).getUser().getId().equals(userDetails.getUser().getId())
+        )) {
+            throw new AccessDeniedException("");
+        }
         commentService.deleteComment(id);
     }
